even

+++ divideelements.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 using namespace std;
 int main(){
-    long long int n;
+    long int n;
     cin >>n;
         if(n%2 ==0){
             cout<<"1";
